Split the native object capture and realize passes by topic
NativeCapture.cs (2184 lines) and NativeRealize.cs (1819) are the two halves of
reproducing an object-tool operation on another machine, and each had grown to
hold every part of that job at once - tool observation, commit matching,
specialized areas, tool input, spawnable tracking, portable references, the
specialized-industry rules, definition building.

Capture becomes eight partial files and realize six, of 150 to 430 lines, each
noting what belongs in it. The two entry files keep the shared state and the
queue-and-retry loop.

Pure code movement, verified member by member; the only textual change is the
blank line each file no longer needs before its first member.
